Fix NUMERIC column type bugs found during Phase 2 activation
- migrations/002: skill column INTEGER -> NUMERIC(4,2) (staff had 0.5 values)
- staffRepo: parseFloat(skill) — pg returns NUMERIC as string
- ipcRepo: parseFloat(overall_score, compliance_pct)
- complaintSurveyRepo: parseFloat(overall_satisfaction)

+++ migrations/002_create_staff.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS staff (
   role            VARCHAR(100)   NOT NULL,
   team            VARCHAR(50)    NOT NULL,
   pref            VARCHAR(10),
-  skill           INTEGER        NOT NULL DEFAULT 1,
+  skill           NUMERIC(4,2)   NOT NULL DEFAULT 1,
   hourly_rate     NUMERIC(8,2)   NOT NULL,
   active          BOOLEAN        NOT NULL DEFAULT true,
   wtr_opt_out     BOOLEAN        NOT NULL DEFAULT false,
